Record L3 source files and export commands
These are the files that can be used to build most of
the autogenerated files in the directory at

  examples/l3-machine-code/arm8/model

The arm8.{sig,sml} files are not currently generated,
and the arm8AssemblerLib.{sig,sml} files are manually
constructed.

Files are taken from the isa-models.tar.bz2 file (dated
2019-11-04), and from direct e-mail correspondence with
Anthoy Fox.